+++ src/components/TagsField/TagsField.tsx
@@ -184,7 +184,7 @@ const getStyles = (theme: GrafanaTheme2, placeholder: string): EditorStyles => {
   return {
     queryField: css`
       height: 32px;
-      border-radius: ${theme.shape.borderRadius()};
+      border-radius: ${theme.shape.radius.pill};
       border: 1px solid ${theme.components.input.borderColor};
       flex: 1;
     `,